refactor(settings): 設定ファイル手続きを settings_store service へ吸い上げ (#782)
commands/settings.rs (823 行) から検証・atomic 書込・export/import バンドル
収集を src-tauri/src/settings_store.rs へ移し、コマンドハンドラは
「AppHandle のパス解決 + ダイアログ + OS 統合 (WSL エディタ委譲)」のみに縮退。

- service は base_dir 引数で AppHandle 非依存 — temp dir で直接テストできる。
  従来テスト側に 3 回再実装されていた export/import ロジックの重複を解消し、
  bundle round-trip / traversal 拒否 / allowlist 検証を実物の service 関数で検証
- write_notedeck_json (settings.json5) をほかと同じ atomic write に統一
  (#719 の趣旨に対しここだけ fs::write のままだった不整合を修正)
